requirements.txt
requests>=2.32.0
python-dotenv>=1.0.1
py-clob-client>=0.25.0
websocket-client>=1.8.0
flask>=3.0.0
psycopg[binary,pool]>=3.2.9
